Move servicesSelector to new services folder
