+++ src/client/src/store/userSelection/services/servicesSelector.ts
@@ -1,8 +1,8 @@
 import { createSelector } from "reselect";
-import { AppState } from "../combineReducers";
-import { ServiceState } from "../userSelection/services/combineReducers";
-import { isAppServiceSelectedSelector } from "./appService/selector";
-import { isCosmosResourceCreatedSelector } from "./cosmosDb/selector";
+import { AppState } from "../../combineReducers";
+import { ServiceState } from "./combineReducers";
+import { isAppServiceSelectedSelector } from "../../azureProfileData/appService/selector";
+import { isCosmosResourceCreatedSelector } from "../../azureProfileData/cosmosDb/selector";
 
 const getState = (state: AppState): AppState => state;
